Refactor: Metrics-related code now on `metrics` app ...
This commit moves the all the serializers, views and routers
for the metrics-related endpoint to the dedicated metrics app.

These changes heavily improve maintainability and scalability.
