+++ metrics/views/trades.py
@@ -6,18 +6,13 @@ from rest_framework.permissions import IsAuthenticated
 from trades.models import Trade
 from trades.filters import TradeFilterBackend, TradeFilterSet
 
-
-class TradeViewMixin:
-    model = Trade
-    queryset = model.objects.all()
-    permission_classes = [IsAuthenticated]
-    filter_backends = [TradeFilterBackend]
-    filterset_class = TradeFilterSet
-
-    def get_queryset(self):
-        return super().get_queryset().filter(
-            user=self.request.user
-        )
+from metrics.serializers.trades import (
+    MetricsSummarySerializer,
+    ProfitAndLossSerializer,
+    TotalTradesSerializer,
+    HoldingTimeSerializer,
+    PositionVolumeSerializer,
+)
 
 
 class MetricsViewMixin(GenericAPIView):
@@ -40,3 +35,23 @@ class MetricsViewMixin(GenericAPIView):
         )
         serializer.is_valid(raise_exception=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+class MetricsSummaryView(MetricsViewMixin):
+    serializer_class = MetricsSummarySerializer
+
+
+class ProfitAndLossView(MetricsViewMixin):
+    serializer_class = ProfitAndLossSerializer
+
+
+class TotalTradesView(MetricsViewMixin):
+    serializer_class = TotalTradesSerializer
+
+
+class HoldingTimeView(MetricsViewMixin):
+    serializer_class = HoldingTimeSerializer
+
+
+class PositionVolumeView(MetricsViewMixin):
+    serializer_class = PositionVolumeSerializer
