+++ trades/views.py
@@ -2,16 +2,24 @@ from rest_framework.generics import (
     ListCreateAPIView,
     RetrieveUpdateDestroyAPIView,
 )
+from rest_framework.permissions import IsAuthenticated
 
-from trades.serializers import (
-    TradeSerializer,
-    MetricsSummarySerializer,
-    ProfitAndLossSerializer,
-    TotalTradesSerializer,
-    HoldingTimeSerializer,
-    PositionVolumeSerializer,
-)
-from trades.view_mixins import TradeViewMixin, MetricsViewMixin
+from trades.models import Trade
+from trades.serializers import TradeSerializer
+from trades.filters import TradeFilterBackend, TradeFilterSet
+
+
+class TradeViewMixin:
+    model = Trade
+    queryset = model.objects.all()
+    permission_classes = [IsAuthenticated]
+    filter_backends = [TradeFilterBackend]
+    filterset_class = TradeFilterSet
+
+    def get_queryset(self):
+        return super().get_queryset().filter(
+            user=self.request.user
+        )
 
 
 class TradeListView(TradeViewMixin, ListCreateAPIView):
@@ -20,23 +28,3 @@ class TradeListView(TradeViewMixin, ListCreateAPIView):
 
 class TradeDetailView(TradeViewMixin, RetrieveUpdateDestroyAPIView):
     serializer_class = TradeSerializer
-
-
-class MetricsSummaryView(MetricsViewMixin):
-    serializer_class = MetricsSummarySerializer
-
-
-class ProfitAndLossView(MetricsViewMixin):
-    serializer_class = ProfitAndLossSerializer
-
-
-class TotalTradesView(MetricsViewMixin):
-    serializer_class = TotalTradesSerializer
-
-
-class HoldingTimeView(MetricsViewMixin):
-    serializer_class = HoldingTimeSerializer
-
-
-class PositionVolumeView(MetricsViewMixin):
-    serializer_class = PositionVolumeSerializer
